+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -13,6 +13,10 @@ Notes](../../RELEASENOTES.md).
 * Update `TraceContextPropagator` to support the W3C randomness flag.
   ([#7301](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7301))
 
+* Fix issue with `Baggage.Current` updates made in a child async flow that
+  could leak back into the parent flow.
+  ([#7191](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7191))
+
 ## 1.15.3
 
 Released 2026-Apr-21